+++ templates/Admin/Screen/Budget/BudgetUpdateScreen.jsx
@@ -371,7 +371,7 @@ export default function BudgetUpdateScreen() {
                     />
                   </div>
                   <div>
-                    {parseFloat(article.quantity) * parseFloat(article.price)}
+                    {parseFloat(article.quantity) * parseFloat(article.price)} €
                   </div>
                   <div>
                     <button className="btn btn-outline-secondary d-inline-flex align-items-center">
@@ -414,7 +414,7 @@ export default function BudgetUpdateScreen() {
                       (sum, article) => sum + article.quantity * article.price,
                       0
                     ) * 0.21
-                  ).toFixed(2)}
+                  ).toFixed(2)} €
                 </div>
                 <div>
                   {values.articles
@@ -422,7 +422,7 @@ export default function BudgetUpdateScreen() {
                       (sum, article) => sum + article.quantity * article.price,
                       0
                     )
-                    .toFixed(2)}
+                    .toFixed(2)} €
                 </div>
                 <div>
                   {(
@@ -433,7 +433,7 @@ export default function BudgetUpdateScreen() {
                           parseFloat(article.price),
                       0
                     ) * 1.21
-                  ).toFixed(2)}
+                  ).toFixed(2)} €
                 </div>
               </ResultTable>
