--- templates/Admin/Screen/Company/CompanyHomeScreen.jsx
@@ -1,9 +1,0 @@
-import CompanyEdit from '../../Layout/Component/Specific/Company/CompanyEdit';
-
-export default function CompanyHomeScreen() {
-  return (
-    <>
-      <CompanyEdit />
-    </>
-  );
-}
